local-resume-parser/requirements.txt
llama-cpp-python==0.2.78
requests==2.32.3
